Revert "Merge branch 'dev-adods-raya' into 'dev-adods'"
This reverts merge request !5

+++ app/Base/BaseRepository.php
@@ -124,7 +124,9 @@ abstract class BaseRepository
      *
      * @return void
      */
-    public function setAutoResourceCreateValidate() {}
+    public function setAutoResourceCreateValidate()
+    {
+    }
 
     /**
      * Override
@@ -133,7 +135,9 @@ abstract class BaseRepository
      *
      * @return void
      */
-    public function setAutoResourceUpdateValidate() {}
+    public function setAutoResourceUpdateValidate()
+    {
+    }
 
     /**
      * Magic function for method with prefix:
@@ -803,7 +807,7 @@ abstract class BaseRepository
                 $idAsKey = true;
                 unset($filter['idAsKey']);
             }
-
+            
             $model = $this->_filter($model, $filter);
             if (isset($filter['hiddenColumn'])) {
                 $hiddenColumn = $filter['hiddenColumn'];
@@ -1007,7 +1011,7 @@ abstract class BaseRepository
      * @param \Illuminate\Database\Eloquent\Model $model
      * @param array $filter
      *      datarule
-     *
+     * 
      * @return \Illuminate\Database\Eloquent\Model
      */
     final protected function _filter($model, array $filter = [])
@@ -1017,13 +1021,13 @@ abstract class BaseRepository
         $qSearch = null;
         $searchField = null;
 
-        if (isset($filter['logquery']))
+        if(isset($filter['logquery']))
             unset($filter['logquery']);
 
         // jika ada datarule
-        if (array_key_exists('datarule', $filter)) {
-            if ($filter['datarule'])
-                $model = $this->handleDatarule($model, $filter['datarule']);
+        if (array_key_exists('datarule',$filter)) {
+            if($filter['datarule'])
+                $model = $this->handleDatarule($model,$filter['datarule']);            
             unset($filter['datarule']);
         }
 
@@ -1084,9 +1088,9 @@ abstract class BaseRepository
         return $model;
     }
 
-    protected function handleDatarule($model, $filterDatrule)
-    {
-        $model = \hpsynapse\moduser\Facades\UserAuth::filterDatarule($model, $filterDatrule);
+    protected function handleDatarule($model,$filterDatrule)
+    {        
+        $model = \hpsynapse\moduser\Facades\UserAuth::filterDatarule($model,$filterDatrule);  
         return $model;
     }
 
@@ -1246,6 +1250,7 @@ abstract class BaseRepository
             } else {
                 $this->error = __('lang.data_not_found');
             }
+
         } catch (\Illuminate\Database\QueryException $ex) {
             $this->error = $ex->getMessage();
         }
@@ -1301,9 +1306,8 @@ abstract class BaseRepository
 
         $model = $this->_where($model, $where);
 
-        // if ($model && $model->exists()) {
-        if ($model) {
-            if (!$model->exists())
+        if ($model && $model->exists()) {
+            if ($model->count() <= 0)
                 return true;
             if ($model->delete())
                 return true;
